Update to 0.18.1722 rev1196

+++ Assembly-CSharp/Verse/ByteGrid.cs
@@ -86,7 +86,7 @@ namespace Verse
 		{
 			if (this.MapSizeMatches(map) && this.grid != null)
 			{
-				this.Clear((byte)0);
+				this.Clear(0);
 			}
 			else
 			{
