+++ Assembly-CSharp/Verse/Printer_Shadow.cs
@@ -4,7 +4,7 @@ namespace Verse
 {
 	public static class Printer_Shadow
 	{
-		private static readonly Color32 LowVertexColor = new Color32((byte)0, (byte)0, (byte)0, (byte)0);
+		private static readonly Color32 LowVertexColor = new Color32(0, 0, 0, 0);
 
 		public static void PrintShadow(SectionLayer layer, Vector3 center, ShadowData shadow, Rot4 rotation)
 		{
@@ -16,7 +16,7 @@ namespace Verse
 			if (DebugViewSettings.drawShadows)
 			{
 				LayerSubMesh subMesh = layer.GetSubMesh(MatBases.SunShadowFade);
-				Color32 item = new Color32((byte)255, (byte)0, (byte)0, (byte)Mathf.Min((float)(255.0 * volume.y), 255f));
+				Color32 item = new Color32(255, 0, 0, (byte)Mathf.Min((float)(255.0 * volume.y), 255f));
 				Vector3 vector = volume.RotatedBy(rotation) / 2f;
 				float x = center.x;
 				float z = center.z;
